gpu: nvgpu: MISRA 10.x fix
-- This will MISRA 10.x violations in semaphore_pool.c,
   nvgpu_mem.h, nvgpu_mem.c and posix-nvgpu_mem.c.

JIRA NVGPU-3177

Reviewed-on: https://git-master.nvidia.com/r/2119524
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/include/nvgpu/nvgpu_mem.h
@@ -220,7 +220,7 @@ static inline bool nvgpu_mem_is_valid(struct nvgpu_mem *mem)
  */
 int nvgpu_mem_create_from_mem(struct gk20a *g,
 			      struct nvgpu_mem *dest, struct nvgpu_mem *src,
-			      u64 start_page, int nr_pages);
+			      u64 start_page, size_t nr_pages);
 
 /**
  * nvgpu_mem_create_from_phys - Create an nvgpu_mem from physical mem.
